add comment

+++ src/main/java/cn/edu/tit/forum/dto/QuestionDTO.java
@@ -11,12 +11,12 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creater;
+    private Long creater;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
